+++ DORM/Includes/Abstracts/JobAfter.php
@@ -1,0 +1,7 @@
+<?php
+namespace DORM\Includes\Abstracts;
+
+abstract class JobAfter {
+
+	abstract static public function do($after, &$job, &$result, &$errors): void;
+}